pyproject.toml
[project]
name = "toolcalling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
